+++ src/main/java/org/jukeboxmc/block/BlockSlab.java
@@ -46,7 +46,7 @@ public abstract class BlockSlab extends BlockWaterlogable {
     }
 
     public BlockSlab setTopSlot( boolean value ) {
-        this.setState( "top_slot_bit", value ? (byte) 1 : (byte) 0 );
+        this.setState( "top_slot_bit", value ? 1 : 0 );
         return this;
     }
